Upgrade form Tauri 1 to 2
Upgrade from Tauri 1 to 2

Cleanup
