+++ src-tauri/Cargo.toml
@@ -1,27 +1,35 @@
 [package]
 name = "game-launcher"
-version = "0.0.2"
+version = "0.0.3"
 description = "Game Launcher"
 authors = ["you"]
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+# The `_lib` suffix may seem redundant but it is necessary
+# to make the lib name unique and wouldn't conflict with the bin name.
+# This seems to be only an issue on Windows, see https://github.com/rust-lang/cargo/issues/8519
+name = "game_launcher_lib"
+crate-type = ["staticlib", "cdylib", "rlib"]
+
 [build-dependencies]
-tauri-build = { version = "1", features = [] }
+tauri-build = { version = "2", features = [] }
 
 [dependencies]
-tauri = { version = "1", features = [ "dialog-open", "updater", "http-all", "fs-all", "shell-open"] }
-tauri-plugin-store = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
+tauri = { version = "2", features = [] }
+tauri-plugin-shell = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
-reqwest = { version = "0.11.26", features = ["blocking"] }
-futures-util = "0.3.30"
-tokio = "1.36.0"
-tokio-stream = "0.1.15"
-futures = "0.3.30"
+tauri-plugin-http = "2"
+tauri-plugin-dialog = "2"
+tauri-plugin-store = "2"
+reqwest = "0.12.9"
+tokio = "1.42.0"
 sha2 = "0.10.8"
+futures-util = "0.3.31"
 
-[features]
-# This feature is used for production builds or when a dev server is not specified, DO NOT REMOVE!!
-custom-protocol = ["tauri/custom-protocol"]
+[target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
+tauri-plugin-updater = "2"
+
