+++ src-tauri/tauri.conf.json
@@ -1,36 +1,15 @@
 {
+  "$schema": "https://schema.tauri.app/config/2",
+  "productName": "game-launcher",
+  "version": "0.0.3",
+  "identifier": "com.estheban.game-launcher.app",
   "build": {
-    "beforeDevCommand": "npm run dev",
-    "beforeBuildCommand": "npm run build",
-    "devPath": "http://localhost:1420",
-    "distDir": "../dist"
+    "beforeDevCommand": "pnpm dev",
+    "devUrl": "http://localhost:1420",
+    "beforeBuildCommand": "pnpm build",
+    "frontendDist": "../dist"
   },
-  "package": {
-    "productName": "game-launcher"
-  },
-  "tauri": {
-    "allowlist": {
-      "all": false,
-      "dialog": {
-        "open": true
-      },
-      "shell": {
-        "all": false,
-        "open": true
-      },
-      "fs": {
-        "all": true,
-        "scope": ["**"]
-      },
-      "http": {
-        "all": true,
-        "request": true,
-        "scope": [
-          "https://yulbrew-game-launcher-dev.s3.ca-central-1.amazonaws.com/*",
-          "http://127.0.0.1:8000/*"
-        ]
-      }
-    },
+  "app": {
     "windows": [
       {
         "title": "game-launcher",
@@ -40,19 +19,21 @@
     ],
     "security": {
       "csp": null
-    },
-    "bundle": {
-      "active": true,
-      "targets": "all",
-      "identifier": "com.estheban.dev.game-launcher",
-      "icon": [
-        "icons/32x32.png",
-        "icons/128x128.png",
-        "icons/128x128@2x.png",
-        "icons/icon.icns",
-        "icons/icon.ico"
-      ]
-    },
+    }
+  },
+  "bundle": {
+    "active": true,
+    "targets": "all",
+    "icon": [
+      "icons/32x32.png",
+      "icons/128x128.png",
+      "icons/128x128@2x.png",
+      "icons/icon.icns",
+      "icons/icon.ico"
+    ],
+    "createUpdaterArtifacts": "v1Compatible"
+  },
+  "plugins": {
     "updater": {
       "active": true,
       "endpoints": [
