+++ src/services/GameManifestService.ts
@@ -1,8 +1,7 @@
 
-import { fetch, ResponseType } from '@tauri-apps/api/http';
+import { fetch } from '@tauri-apps/plugin-http';
 import {Manifest} from "../types/Manifest.ts";
-import {invoke} from "@tauri-apps/api/tauri";
-// import {invoke} from "@tauri-apps/api/tauri";
+import { invoke } from "@tauri-apps/api/core";
 
 export default class GameManifestService {
     async get_download_prefix(): Promise<string> {
@@ -10,12 +9,17 @@ export default class GameManifestService {
     }
 
     async get(): Promise<Manifest> {
+        // @todo: handle errors
         // const response = await fetch('http://127.0.0.1:8000/games/73dd1271-d2d9-4db6-9618-13ddec1a073b', {
         const response = await fetch(await this.get_download_prefix() + 'manifest.json', {
             method: 'GET',
             timeout: 10,
-            responseType: ResponseType.JSON,
+            // responseType: ResponseType.JSON,
         });
-        return await response.data as Manifest;
+        console.log(response.status); // e.g. 200
+
+        const jsonData: Manifest = await response.json();
+
+        return jsonData;
     }
 }
